Advanced Decor Deleting
Deleting decor that crosses chunk boundaries now completely deletes all other fragments of the split of decor.

+++ scripts/input.py
@@ -30,7 +30,7 @@ class Input:
         mx = (self.mouseposition[0] - self.editor.window.toolbar.width) * self.editor.window.camera.ratio[0]
         my = self.mouseposition[1] * self.editor.window.camera.ratio[1]
         if self.currentAssetType == 'decorations':
-            return mx * self.editor.window.camera.zoom + self.editor.window.camera.scroll[0], my * self.editor.window.camera.zoom + self.editor.window.camera.scroll[1]
+            return int(mx * self.editor.window.camera.zoom + self.editor.window.camera.scroll[0]), int(my * self.editor.window.camera.zoom + self.editor.window.camera.scroll[1])
         elif self.currentAssetType == 'tiles':
             return (mx * self.editor.window.camera.zoom + self.editor.window.camera.scroll[0]) // self.TILESIZE, (my * self.editor.window.camera.zoom + self.editor.window.camera.scroll[1]) // self.TILESIZE
